pyproject.toml
[project]
name = "stylish-tts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "numba>=0.60.0",
    "numpy>=2.0.2",
    "munch>=4.0.0",
    "einops>=0.8.0",
    "einops-exts>=0.0.4",
    "espeakng-loader>=0.2.4",
    "phonemizer-fork>=3.3.2",
    "transformers>=4.48.2",
    "monotonic-align>=1.0.0",
    "matplotlib>=3.10.0",
    "scikit-learn>=1.6.1",
    "librosa>=0.10.2.post1",
    "nnaudio>=0.3.3",
    "accelerate>=1.3.0",
    "pydantic>=2.10.6",
    "tensorboard>=2.18.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
]

[tool.uv.sources]
monotonic-align = { git = "https://github.com/resemble-ai/monotonic_align.git" }
